pyproject.toml
[project]
name = "fastapi-ai-pet-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]
